implement PR suggestions by Melvin
