+++ database/seed-database.test.js
@@ -1,5 +1,6 @@
 const mongoose = require('mongoose');
 const {Business} = require('./db-helpers');
+const { allFields, hasRequisiteFields } = require('../test/helpers.js');
 
 beforeAll( async () => {
   await mongoose.connect('mongodb://localhost/businessInfo');
@@ -10,10 +11,11 @@ afterAll( async () => {
 })
 
 describe('should have seeded data in database', () => {
-  test('should have data in database', done => {
+  test('should have data with correct fields in database', done => {
     Business.find({id: 201}).exec( (err, result) => {
       if (err) { console.log(err); }
       expect(result).toBeDefined();
+      hasRequisiteFields(allFields, result[0]);
       done();
     });
   });
